Split handout text chunks and lazy-load search
