+++ client/src/__tests__/search.test.tsx
@@ -2,6 +2,24 @@ import { act, fireEvent, render, screen } from "@testing-library/react";
 import { Router } from "wouter";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import Search, { getSearchTerms } from "@/components/Search";
+import { searchableContent } from "@/content/searchIndex";
+
+const { loadNormalizedSearchIndex } = vi.hoisted(() => ({
+  loadNormalizedSearchIndex: vi.fn(),
+}));
+
+vi.mock("@/content/searchIndexLoader", () => ({
+  loadNormalizedSearchIndex,
+  resetSearchIndexCacheForTests: vi.fn(),
+}));
+
+const normalizedSearchEntries = searchableContent.map(item => ({
+  item,
+  normalizedTitle: item.title.toLowerCase(),
+  searchText: [item.title, item.description, ...item.keywords, item.section]
+    .join(" ")
+    .toLowerCase(),
+}));
 
 describe("Search helpers", () => {
   it("ignores whitespace-only and one-letter search tokens", () => {
@@ -24,6 +42,8 @@ describe("Search", () => {
       writable: true,
     });
     vi.useFakeTimers();
+    loadNormalizedSearchIndex.mockReset();
+    loadNormalizedSearchIndex.mockResolvedValue(normalizedSearchEntries);
   });
 
   afterEach(() => {
@@ -31,7 +51,7 @@ describe("Search", () => {
     vi.useRealTimers();
   });
 
-  it("clears the active descendant when the query is no longer searchable", () => {
+  it("clears the active descendant when the query is no longer searchable", async () => {
     render(
       <Router>
         <Search isOpen onClose={() => {}} />
@@ -42,11 +62,17 @@ describe("Search", () => {
       name: /website durchsuchen/i,
     });
 
+    await act(async () => {
+      await Promise.resolve();
+    });
+
     fireEvent.change(input, { target: { value: "Validierung" } });
 
     act(() => {
       vi.advanceTimersByTime(200);
     });
+
+    expect(screen.getAllByRole("option").length).toBeGreaterThan(0);
 
     fireEvent.keyDown(input, { key: "ArrowDown" });
 
@@ -58,5 +84,29 @@ describe("Search", () => {
     expect(
       screen.getByText(/mindestens einen suchbegriff mit 2 zeichen/i)
     ).toBeInTheDocument();
+  });
+
+  it("shows a loading state while the search index is still loading", async () => {
+    loadNormalizedSearchIndex.mockReturnValueOnce(
+      new Promise(() => {
+        /* keep pending to assert the loading UI */
+      })
+    );
+
+    await act(async () => {
+      render(
+        <Router>
+          <Search isOpen onClose={() => {}} />
+        </Router>
+      );
+    });
+
+    await act(async () => {
+      fireEvent.change(screen.getByRole("combobox"), {
+        target: { value: "Krise" },
+      });
+    });
+
+    expect(screen.getByText(/suchindex wird geladen/i)).toBeInTheDocument();
   });
 });
