+++ client/src/content/handoutTextVersionContent/shared.ts
@@ -1,0 +1,15 @@
+import { createHandoutTextVersionMeta } from "../handoutTextVersionRegistry";
+import type {
+  HandoutTextVersion,
+  HandoutTextVersionContent,
+} from "../handoutTextVersionTypes";
+
+export function createHandoutTextVersion(
+  id: string,
+  config: HandoutTextVersionContent
+): HandoutTextVersion {
+  return {
+    ...createHandoutTextVersionMeta(id),
+    ...config,
+  };
+}
